requirements.txt
Flask==2.2.3
flask_basicauth==0.2.0
humanize==4.6.0
loguru==0.6.0
mysql-connector==2.2.9
